Add enzyme to this unit test
